requirements.txt
requests
beautifulsoup4
pandas
python-telegram-bot
millify
matplotlib
